Finish all endpoint in student controller

+++ src/main/java/com/rith/_15_soeun_sovannarith_pvh_spring_homework02/models/Entity/Instructor.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 public class Instructor {
 
-    private Long instructorId;
+    private Integer instructorId;
     private String instructorName;
     private String email;
 }
